changed mode: SmallPositiveIntegert --> PositiveInteger

+++ api/models.py
@@ -24,7 +24,7 @@ class SatelliteInfo(models.Model):
 
 class SatNameId(models.Model):
     Name = models.CharField(max_length=50)
-    SatId = models.PositiveSmallIntegerField()
+    SatId = models.PositiveIntegerField()
 
     def __str__(self):
         return self.Name
